+++ plaso/storage/zip_file.py
@@ -144,60 +144,6 @@ from plaso.storage import interface
 from plaso.storage import gzip_file
 
 
-class _SerializedEventHeap(object):
-  """Serialized event heap.
-
-  Attributes:
-    data_size (int): total data size of the serialized events on the heap.
-  """
-
-  def __init__(self):
-    """Initializes a serialized event heap."""
-    super(_SerializedEventHeap, self).__init__()
-    self._heap = []
-    self.data_size = 0
-
-  @property
-  def number_of_events(self):
-    """int: number of serialized events on the heap."""
-    return len(self._heap)
-
-  def Empty(self):
-    """Empties the heap."""
-    self._heap = []
-    self.data_size = 0
-
-  def PopEvent(self):
-    """Pops an event from the heap.
-
-    Returns:
-      tuple: contains:
-
-        int: event timestamp or None if the heap is empty
-        bytes: serialized event or None if the heap is empty
-    """
-    try:
-      timestamp, serialized_event = heapq.heappop(self._heap)
-
-      self.data_size -= len(serialized_event)
-      return timestamp, serialized_event
-
-    except IndexError:
-      return None, None
-
-  def PushEvent(self, timestamp, event_data):
-    """Pushes a serialized event onto the heap.
-
-    Args:
-      timestamp (int): event timestamp, which contains the number of
-          micro seconds since January 1, 1970, 00:00:00 UTC.
-      event_data (bytes): serialized event.
-    """
-    heap_values = (timestamp, event_data)
-    heapq.heappush(self._heap, heap_values)
-    self.data_size += len(event_data)
-
-
 class _SerializedDataStream(object):
   """Serialized data stream.
 
@@ -782,7 +728,7 @@ class ZIPStorageFile(interface.BaseStorageFile):
     self._offset_tables_lfu = []
     self._path = None
     self._saved_stream_offsets = {}
-    self._serialized_event_heap = _SerializedEventHeap()
+    self._serialized_event_heap = event_heaps.SerializedEventHeap()
     self._serialized_event_tags = []
     self._serialized_event_tags_size = 0
     self._open_streams = {}
@@ -2044,7 +1990,7 @@ class ZIPStorageFile(interface.BaseStorageFile):
     """Writes the contents of an serialized event heap.
 
     Args:
-      serialized_event_heap(_SerializedEventHeap): serialized event heap.
+      serialized_event_heap(SerializedEventHeap): serialized event heap.
       stream_number(int): stream number.
     """
     stream_name = 'event_index.{0:06d}'.format(stream_number)
